Refactor tests to work async
In theory the whole plugin works with a asynchronous stream
of files. To be a bit mor flexible with refactorings I added
the expected behavior to all tests. So, tests will wait
for any expectations until the stream is finished.

On this changes I noticed that we can run into a runtime problem
between `mvn.config()` and `mvn.deploy()` caused by `async.each`
if any deploy streams will be executed in parallel. So, I moved
the config call to just above the deploy call.